DEmO

+++ admin/app/Http/Controllers/CategoryController.php
@@ -18,51 +18,62 @@ class CategoryController extends Controller
 *
 * @return \Illuminate\Http\Response
 */
-public function index()
-{
-    $category= Category::latest()->paginate();
-    $categories=DB::table('category as c1')
-    ->select('c1.id','c1.name','c1.status','c1.parent_id','c2.name as parent_name')
-    ->join('category as c2','c1.parent_id','=','c2.id')
+    public function index()
+    {
+        $categories=Category::select('category.id','category.name','category.status','category.parent_id','c2.name as parent_name')
+                            ->leftjoin('category as c2','category.parent_id','=','c2.id')
+                            ->paginate(4);
 
-    ->get();
+        return view('category.index',compact('categories','categorys','categorys','category'));
 
-    $categorys=Category::select('*')->where('parent_id','=','0')->get();
-
-
-    return view('category.index',compact('categories','categorys','category','category'));
-
-}
+    }
 
 /**
 * Show the form for creating a new resource.
 *
 * @return \Illuminate\Http\Response
 */
-public function create()
-{
-    $category=Category::all();
-    return view('category.create',compact('category'));
-}
+    public function create()
+    {
+        $category=Category::all();
+        return view('category.create',compact('category'));
+    }
 
+    
+    public function search(Request $request){
+        $search=$request->search;
+        $categories=Category::select('category.id','category.name','category.status','category.parent_id','c2.name as parent_name')
+                            ->leftjoin('category as c2','category.parent_id','=','c2.id')
+                            ->where('category.name','like','%'.$search.'%')
+                            ->paginate(4);
+        
+       
+        return view('category.index',compact('categories'));
+
+    }
 /**
 * Store a newly created resource in storage.
 *
 * @param \Illuminate\Http\Request $request
 * @return \Illuminate\Http\Response
 */
-public function store(Request $request)
-{
-    $request->validate([
+    public function store(Request $request)
+    {
+        $request->validate([
 
-        'name'=>'required',
-        'status'=>'required',
-        'parent_id'=>'required'
-    ]);
-// $category->parent_id=$request->parent_id;
-    Category::create($request->all());
-    return redirect()->route('category.index')->with('success','Category created successfully');
-}
+            'name'=>'required',
+            'status'=>'required',
+            'parent_id'=>'required'
+        ]);
+        $name=strtoupper($request->name);
+        $category=new Category();
+        $category->name=$name;
+        $category->status=$request->status;
+        $category->parent_id=$request->parent_id;
+        $category->save();
+       
+        return redirect()->route('category.index')->with('success','Category created successfully');
+    }
 
 
 /**
@@ -71,10 +82,10 @@ public function store(Request $request)
 * @param int $id
 * @return \Illuminate\Http\Response
 */
-public function show($id)
-{
-//
-}
+    public function show($id)
+    {
+    //
+    }
 
 /**
 * Show the form for editing the specified resource.
@@ -82,16 +93,14 @@ public function show($id)
 * @param int $id
 * @return \Illuminate\Http\Response
 */
-public function edit($id)
-{
-   
-
-    $categories=Category::where('id','=',$id)->first();
-    $category = Category::find($id); 
-    $categoryDetails=Category::where('id',$id)->first();
-    $level=Category::where('parent_id',0)->get();
-    return view('category.edit',compact('category','categories','categoryDetails','level'));
-}
+    public function edit($id)
+    {
+        $categories=Category::where('id','=',$id)->first();
+        $category = Category::find($id); 
+        $categoryDetails=Category::where('id',$id)->first();
+        $level=Category::where('parent_id',0)->get();
+        return view('category.edit',compact('category','categories','categoryDetails','level'));
+    }
 
 /**
 * Update the specified resource in storage.
@@ -100,19 +109,20 @@ public function edit($id)
 * @param int $id
 * @return \Illuminate\Http\Response
 */
-public function update(Request $request, $id)
-{
-    $data=$request->all();
-        
-    $category = Category::find($id);
-    $request->validate([
+    public function update(Request $request, $id)
+    {
+        $data=$request->all();
+            
+        $category = Category::find($id);
+        $request->validate([
 
-        'name'=>'required',
-        'status'=>'required'
-    ]);
-    Category::find($id)->update($request->all());
-    return redirect()->route('category.index')->with('success','Category updated successfully');
-}
+            'name'=>'required',
+            'status'=>'required'
+        ]);
+        $name=strtoupper($request->name);
+        Category::find($id)->update(['name'=>$name,'status'=>$request->status]);
+        return redirect()->route('category.index')->with('success','Category updated successfully');
+    }
 
 /**
 * Remove the specified resource from storage.
@@ -120,24 +130,24 @@ public function update(Request $request, $id)
 * @param int $id
 * @return \Illuminate\Http\Response
 */
-public function destroy($id)
-{
-$category_id=Product_categories::where('category_id',$id)->first();
-$product=Product::where('id',$category_id->product_id)->first();
-$product_attribute_assoc=Product_attributes_assoc::where('product_id',$product->id)->first();
-//$product_attributes=Product_attributes::where('id',$product_attribute_assoc->product_attribute_id)->first();
-//$product_attribute_value=Product_attribute_values::where('product_attribute_id',$product_attributes->id)->first();
-//Product_attribute_values::where('product_attribute_id',$product_attributes->id)->delete();
-//Product_attributes::where('id',$product_attribute_assoc->product_attribute_id)->delete();
-Product_attributes_assoc::where('product_id',$product->id)->delete();
-Product_categories::where('category_id',$id)->delete();
-Product_images::where('product_id',$product->id)->delete();
-Category::where('id',$id)->delete();
-//Product_attributes_assoc::where('product_id',$product->id)->delete();
-Product::where('id',$category_id->product_id)->delete();
+    public function destroy($id)
+    {
+        $category_id=Product_categories::where('category_id',$id)->first();
+        if(empty($category_id)){
+        Product_categories::where('category_id',$id)->delete();
+        Category::where('parent_id',$id)->delete();
+        Category::where('id',$id)->delete();
+        }
+        else{
+        $product=Product::where('id',$category_id->product_id)->first();
+        $product_attribute_assoc=Product_attributes_assoc::where('product_id',$product->id)->first();
+        Product_attributes_assoc::where('product_id',$product->id)->delete();
+        Product_categories::where('category_id',$id)->delete();
+        Product_images::where('product_id',$product->id)->delete();
+        Category::where('id',$id)->delete();
+        Product::where('id',$category_id->product_id)->delete();
+        }
+        return redirect()->route('category.index')->with('success','Category deleted successfully');
+    }
 
-
-
-return redirect()->route('category.index')->with('success','Category deleted successfully');
-}
 }
